{graph,path}: move Weighting and UniformCost to path
