+++ path/bellman_ford_moore.go
@@ -8,18 +8,18 @@ import "github.com/gonum/graph"
 
 // BellmanFordFrom returns a shortest-path tree for a shortest path from u to all nodes in
 // the graph g, or false indicating that a negative cycle exists in the graph. If the graph
-// does not implement graph.Weighter, graph.UniformCost is used.
+// does not implement graph.Weighter, UniformCost is used.
 //
 // The time complexity of BellmanFordFrom is O(|V|.|E|).
 func BellmanFordFrom(u graph.Node, g graph.Graph) (path Shortest, ok bool) {
 	if !g.Has(u) {
 		return Shortest{from: u}, true
 	}
-	var weight graph.Weighting
+	var weight Weighting
 	if wg, ok := g.(graph.Weighter); ok {
 		weight = wg.Weight
 	} else {
-		weight = graph.UniformCost(g)
+		weight = UniformCost(g)
 	}
 
 	nodes := g.Nodes()
